refactor(dashboard): change WeekNumber type from long to double

+++ PsychologySupport/Services/ChatBox/ChatBox.API/Models/Views/WeeklyRetentionStat.cs
@@ -2,7 +2,7 @@
 
 public class WeeklyRetentionStat
 {
-    public long WeekNumber { get; set; } // Tuần thứ 1, 2, 3...
+    public double WeekNumber { get; set; } // Tuần thứ 1, 2, 3...
     public long EligibleUsers { get; set; } // Tổng user đủ thâm niên để đạt mốc này
     public long RetainedUsers { get; set; } // Số user thực sự quay lại
     public decimal RetentionRatePercent { get; set; } // Tỷ lệ %
